+++ src/erasure-code/clay/ErasureCodeClay.cc
@@ -95,7 +95,7 @@ unsigned int ErasureCodeClay::get_chunk_size(unsigned int stripe_width) const
   return round_up_to(stripe_width, alignment) / k;
 }
 
-unsigned int ErasureCodeClay::get_minimum_granularity()
+size_t ErasureCodeClay::get_minimum_granularity()
 {
   return mds.erasure_code->get_minimum_granularity();
 }
